+++ src/Controller/Frontend/DefaultController.php
@@ -2,7 +2,6 @@
 
 namespace App\Controller\Frontend;
 
-use App\Entity\Archive;
 use App\Entity\MenuLevel1;
 use App\Entity\MenuLevel2;
 use App\Entity\Page;
@@ -116,45 +115,6 @@ final class DefaultController extends AbstractController
             [
                 'menu' => $menu,
                 'submenu' => $submenu,
-                'show_debug_page_info' => $kernel->isDebug(),
-            ]
-        );
-    }
-
-    /**
-     * @Route({"ca": "/{menu}/any/{year}", "es": "/{menu}/ano/{year}"}, name="front_app_archive_year_list")
-     * @ParamConverter("menu", class="App\Entity\MenuLevel1", options={"mapping": {"menu": "slug"}})
-     * @ParamConverter("archive", class="App\Entity\Archive", options={"mapping": {"year": "year"}})
-     */
-    public function archiveYearList(MenuLevel1 $menu, Archive $archive, PageRepository $pr, KernelInterface $kernel): Response
-    {
-        $pages = $pr->getActiveItemsFromArchive($archive)->getQuery()->getResult();
-
-        return $this->render(
-            'frontend/archive/archive_year_list.html.twig',
-            [
-                'menu' => $menu,
-                'archive' => $archive,
-                'pages' => $pages,
-                'show_debug_page_info' => $kernel->isDebug(),
-            ]
-        );
-    }
-
-    /**
-     * @Route({"ca": "/{menu}/any/{year}/pagina/{slug}/{id}", "es": "/{menu}/año/{year}/pagina/{slug}/{id}"}, name="front_app_archive_year_page_detail")
-     * @ParamConverter("menu", class="App\Entity\MenuLevel1", options={"mapping": {"menu": "slug"}})
-     * @ParamConverter("archive", class="App\Entity\Archive", options={"mapping": {"year": "year"}})
-     * @ParamConverter("page", class="App\Entity\Page", options={"mapping": {"id": "id"}})
-     */
-    public function archiveYearPageDetail(MenuLevel1 $menu, Archive $archive, Page $page, PageRepository $pr, KernelInterface $kernel): Response
-    {
-        return $this->render(
-            'frontend/archive/archive_year_page_detail.html.twig',
-            [
-                'menu' => $menu,
-                'archive' => $archive,
-                'page' => $page,
                 'show_debug_page_info' => $kernel->isDebug(),
             ]
         );