+++ CHANGELOG.md
@@ -5,6 +5,8 @@ Changelog
  * fix bugs
  * fix broken testing suite
  * add page template types
+ * add archive frontend views
+ * add irradiador frontend views
 
 ##### Version 5.07.00 (2021-11-19)
  * fix missing assets install command execution during Ansible deploy